Refactor Time.h and Time.cpp to GenTime.h and GenTime.cpp to avoid conflict with standard template library time.h. Fixed the evaluation of a complex number with the function fabs in Diagnostics.cpp
