+++ .idea/workspace.xml
@@ -42,7 +42,22 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="69343f7c-ea14-439e-8472-60a88f1f7a2d" name="Default Changelist" comment="Update of CHANGELOG.md">
-      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/CMakeLists.txt" beforeDir="false" afterPath="$PROJECT_DIR$/CMakeLists.txt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/include/AlterSetup.h" beforeDir="false" afterPath="$PROJECT_DIR$/include/AlterSetup.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/include/EField.h" beforeDir="false" afterPath="$PROJECT_DIR$/include/EField.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/include/ImportBeam.h" beforeDir="false" afterPath="$PROJECT_DIR$/include/ImportBeam.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/include/ImportField.h" beforeDir="false" afterPath="$PROJECT_DIR$/include/ImportField.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/include/LoadBeam.h" beforeDir="false" afterPath="$PROJECT_DIR$/include/LoadBeam.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/include/LoadField.h" beforeDir="false" afterPath="$PROJECT_DIR$/include/LoadField.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/include/SDDSBeam.h" beforeDir="false" afterPath="$PROJECT_DIR$/include/SDDSBeam.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/include/SponRad.h" beforeDir="false" afterPath="$PROJECT_DIR$/include/SponRad.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/include/Time.h" beforeDir="false" afterPath="$PROJECT_DIR$/include/GenTime.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/include/Track.h" beforeDir="false" afterPath="$PROJECT_DIR$/include/Track.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/include/Wake.h" beforeDir="false" afterPath="$PROJECT_DIR$/include/Wake.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/include/readBeamHDF5.h" beforeDir="false" afterPath="$PROJECT_DIR$/include/readBeamHDF5.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/include/readFieldHDF5.h" beforeDir="false" afterPath="$PROJECT_DIR$/include/readFieldHDF5.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/Main/GenMain.cpp" beforeDir="false" afterPath="$PROJECT_DIR$/src/Main/GenMain.cpp" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/Main/Time.cpp" beforeDir="false" afterPath="$PROJECT_DIR$/src/Main/GenTime.cpp" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -52,7 +67,7 @@
   <component name="ClangdSettings">
     <option name="formatViaClangd" value="false" />
   </component>
-  <component name="ExecutionTargetManager" SELECTED_TARGET="CMakeBuildProfile:Release" />
+  <component name="ExecutionTargetManager" SELECTED_TARGET="CMakeBuildProfile:Debug" />
   <component name="Git.Settings">
     <option name="RECENT_BRANCH_BY_REPOSITORY">
       <map>
