+++ include/LoadField.h
@@ -12,7 +12,7 @@
 
 #include "StringProcessing.h"
 #include "Setup.h"
-#include "Time.h"
+#include "GenTime.h"
 #include "Profile.h"
 #include "GaussHermite.h"
 #include "Field.h"
