+++ include/readFieldHDF5.h
@@ -12,7 +12,7 @@
 #include <hdf5.h>
 #include "HDF5base.h"
 #include "Setup.h"
-#include "Time.h"
+#include "GenTime.h"
 
 using namespace std;
