+++ src/Core/Diagnostic.cpp
@@ -324,13 +324,13 @@ void DiagBeam::getValues(Beam *beam,std::map<std::string,std::vector<double> >&v
             if (val.find("pyposition") != val.end()){ val["pyposition"][idx]=py1;}
         }
 
-        if (val.find("bunching") != val.end()) {val["bunching"][idx] =fabs(b[0]);}
+        if (val.find("bunching") != val.end()) {val["bunching"][idx] =std::abs(b[0]);}
         if (val.find("bunchingphase") != val.end()) {val["bunchingphase"][idx] =atan2(b[0].imag(),b[0].real());}
 
         char buff[100];
         for (int iharm = 1; iharm < nharm; iharm++) {
             snprintf(buff, sizeof(buff), "bunching%d", iharm + 1);
-            if (val.find(buff) != val.end()) { val[buff][idx] = abs(b[iharm]); }
+            if (val.find(buff) != val.end()) { val[buff][idx] = std::abs(b[iharm]); }
             snprintf(buff, sizeof(buff), "bunchingphase%d", iharm + 1);
             if (val.find(buff) != val.end()) { val[buff][idx] = atan2(b[iharm].imag(), b[iharm].real()); }
         }
